// src/generated/models/workspace-membership-model.ts
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const WorkspaceMembershipActiveToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive',
  'Suspended': 'Suspended'
} as const;
export type WorkspaceMembershipActive = keyof typeof WorkspaceMembershipActiveToLabel;

export const WorkspaceMembershipRoleKeyToLabel = {
  'Owner': 'Owner',
  'Administrator': 'Administrator',
  'Manager': 'Manager',
  'Agent': 'Agent',
  'Approver': 'Approver',
  'KnowledgePublisher': 'Knowledge Publisher',
  'Auditor': 'Auditor',
  'Requester': 'Requester'
} as const;
export type WorkspaceMembershipRoleKey = keyof typeof WorkspaceMembershipRoleKeyToLabel;

export interface WorkspaceMembership {
  /**
   * @displayName Workspace Membership ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Membership Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  membershipLabel: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  active: WorkspaceMembershipActive;
  /**
   * @displayName AddedBy
   */
  addedBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName CreatedAt
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName EndDate
   */
  endDate?: string;
  /**
   * @displayName IsDefaultWorkspace
   * @validationRule Required for create/update operations
   */
  isDefaultWorkspace: boolean;
  /**
   * @displayName PersonId
   * @validationRule Required for create/update operations
   */
  person: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName MembershipType
   * @validationRule Required for create/update operations
   */
  roleKey: WorkspaceMembershipRoleKey;
  /**
   * @displayName StartDate
   * @validationRule Required for create/update operations
   */
  startDate: string;
  /**
   * @displayName WorkspaceId
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _WorkspaceMembership = 'WorkspaceMembership' as const;