// src/generated/models/service-membership-model.ts
import type { Person } from './person-model';
import type { Service } from './service-model';
import type { Workspace } from './workspace-model';

export const ServiceMembershipRoleKeyToLabel = {
  'Agent': 'Agent',
  'ServiceManager': 'ServiceManager',
  'ServiceAdministrator': 'ServiceAdministrator',
  'Approver': 'Approver',
  'Auditor': 'Auditor'
} as const;
export type ServiceMembershipRoleKey = keyof typeof ServiceMembershipRoleKeyToLabel;

export const ServiceMembershipStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive',
  'Suspended': 'Suspended'
} as const;
export type ServiceMembershipStatusKey = keyof typeof ServiceMembershipStatusKeyToLabel;

export interface ServiceMembership {
  /**
   * @displayName Service Membership ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Membership Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  membershipLabel: string;
  /**
   * @displayName Added By
   */
  addedBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Created At
   * @validationRule Required for create/update operations
   */
  createdAt: string;
  /**
   * @displayName End Date
   */
  endDate?: string;
  /**
   * @displayName Person
   * @validationRule Required for create/update operations
   */
  person: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Role Key
   * @validationRule Required for create/update operations
   */
  roleKey: ServiceMembershipRoleKey;
  /**
   * @displayName Service
   * @validationRule Required for create/update operations
   */
  service: Pick<Service, 'id' | 'serviceName'>;
  /**
   * @displayName Service Code
   * @validationRule Required for create/update operations
   */
  serviceCode: string;
  /**
   * @displayName Start Date
   * @validationRule Required for create/update operations
   */
  startDate: string;
  /**
   * @displayName Status Key
   * @validationRule Required for create/update operations
   */
  statusKey: ServiceMembershipStatusKey;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _ServiceMembership = 'ServiceMembership' as const;