// src/generated/models/workspace-module-model.ts
import type { Person } from './person-model';
import type { Module_1 } from './module-1-model';
import type { Workspace } from './workspace-model';

export const WorkspaceModuleConfigurationStatusKeyToLabel = {
  'Configured': 'Configured',
  'Draft': 'Draft',
  'Disabled': 'Disabled'
} as const;
export type WorkspaceModuleConfigurationStatusKey = keyof typeof WorkspaceModuleConfigurationStatusKeyToLabel;

export interface WorkspaceModule {
  /**
   * @displayName Workspace Module ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName WorkspaceModuleName
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  workspaceModuleName: string;
  /**
   * @displayName ConfigurationStatus
   * @validationRule Required for create/update operations
   */
  configurationStatusKey: WorkspaceModuleConfigurationStatusKey;
  /**
   * @displayName DisplayNameOverride
   */
  displayNameOverride?: string;
  /**
   * @displayName IsEnabled
   * @validationRule Required for create/update operations
   */
  enabled: boolean;
  /**
   * @displayName EnabledAt
   */
  enabledAt?: string;
  /**
   * @displayName EnabledBy
   */
  enabledBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName ModuleId
   * @validationRule Required for create/update operations
   */
  moduleDefinition: Pick<Module_1, 'id' | 'moduleName'>;
  /**
   * @displayName SortOrderOverride
   */
  sortOrder?: number;
  /**
   * @displayName WorkspaceId
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _WorkspaceModule = 'WorkspaceModule' as const;